Move getCustomChannels to reposync.common
