+++ backend/satellite_tools/reposync/__init__.py
@@ -2,7 +2,7 @@ import os
 
 if os.environ.get('AS_SUSE', False):
     from suse import (
-        RepoSync, ChannelException, ChannelTimeoutException
+        RepoSync, ChannelException, ChannelTimeoutException, getCustomChannels
     )
 else:
-    from reposync import RepoSync, ChannelException, ChannelTimeoutException
+    from reposync import RepoSync, getCustomChannels
